+++ src/core_modules/capture-core/components/WidgetRelatedStages/hooks/useCanAddNewEventToStage.js
@@ -1,0 +1,8 @@
+import { useMemo } from 'react';
+
+export const useCanAddNewEventToStage = (programStage, existingRelatedEvents) => useMemo(() =>
+    (programStage && existingRelatedEvents
+        ? programStage.repeatable || (!programStage.repeatable && existingRelatedEvents.length === 0)
+        : false),
+[programStage, existingRelatedEvents],
+);