+++ src/core_modules/capture-core/trackedEntityInstances/getSubValues.js
@@ -1,38 +1,38 @@
 // @flow
-import log from 'loglevel';
 import isDefined from 'd2-utilizr/lib/isDefined';
-import { errorCreator } from 'capture-core-utils';
+import { featureAvailable, FEATURES } from 'capture-core-utils';
 import { type DataElement, dataElementTypes } from '../metaData';
 import type { QuerySingleResource } from '../utils/api/api.types';
 
-const GET_SUBVALUE_ERROR = 'Could not get subvalue';
-
 const subValueGetterByElementType = {
     [dataElementTypes.IMAGE]: ({
-        value,
         teiId,
         attributeId,
         absoluteApiPath,
-        querySingleResource,
+        programId,
     }: {
-        value: any,
         teiId: string,
         attributeId: string,
         absoluteApiPath: string,
-        querySingleResource: QuerySingleResource,
-    }) =>
-        querySingleResource({ resource: `fileResources/${value}` })
-            .then(res =>
-                ({
-                    name: res.name,
-                    value: res.id,
-                    url: `${absoluteApiPath}/trackedEntityInstances/${teiId}/${attributeId}/image`,
-                }))
-            .catch((error) => {
-                log.warn(errorCreator(GET_SUBVALUE_ERROR)({ value, teiId, attributeId, error }));
-                return null;
-            }) };
+        programId: ?string,
+    }) => {
+        const buildUrl = () => {
+            if (featureAvailable(FEATURES.trackerImageEndpoint)) {
+                if (programId) {
+                    return `${absoluteApiPath}/tracker/trackedEntities/${teiId}/attributes/${attributeId}/image?program=${programId}&dimension=small`;
+                }
+                return `${absoluteApiPath}/tracker/trackedEntities/${teiId}/attributes/${attributeId}/image?dimension=small`;
+            }
+            return `${absoluteApiPath}/trackedEntityInstances/${teiId}/${attributeId}/image`;
+        };
+        const previewUrl = buildUrl();
 
+        return {
+            previewUrl,
+            url: previewUrl,
+        };
+    },
+};
 
 export async function getSubValues({
     teiId,
@@ -40,12 +40,14 @@ export async function getSubValues({
     values,
     absoluteApiPath,
     querySingleResource,
+    programId,
 }: {
     teiId: string,
     attributes: Array<DataElement>,
     values?: ?Object,
     absoluteApiPath: string,
     querySingleResource: QuerySingleResource,
+    programId: ?string,
 }) {
     if (!values) {
         return null;
@@ -67,6 +69,7 @@ export async function getSubValues({
                     attributeId,
                     absoluteApiPath,
                     querySingleResource,
+                    programId,
                 });
                 accValues[attributeId] = subValue;
             }